pin all versions
We dont need to pin the patch level and I am torn about doing it in
the first place but I did it here for the sake of it. We can pull it
back to just pinning the major and minor versions if needed.

The official pipenv docs say you shouldn;t pin but I would argue
very strongly that you should always pin at least the major and minor
versions of any dependencies, simply for the sake of a consistent
environment if nothing else.