bring in invoke task for initializing the db for testing
I create an invoke task `invoke prep` that will initialize the db for
the user who wishes to simply git clone the repo and run tests. There
is no need to do anything except install the deps with pipenv and then
use the above command.

Namespaces still need to be created and I need to add the ability to
bring in options for dev's that want to put the db and config file
locally instead of the native place and risk messing with their
production install.

I don't see a need for the average user to use invoke but this is more
to make things easier on the dev side and with Travis.